backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.8
alembic==1.12.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.4
pydantic-settings==2.2.1
httpx==0.25.2
redis==5.0.1
celery==5.3.4
email-validator==2.1.0.post1
python-multipart==0.0.9

# ML/NLP libraries will be added in later phases
# tensorflow==2.14.0
# scikit-learn==1.3.2
# pandas==2.1.3
# numpy==1.25.2
# prophet==1.1.5
# spacy==3.7.2
# transformers==4.35.2
